+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-﻿## Version 4.13 [13-05-2026]
+﻿## Version 4.14 [22-05-2026]
+
+**All WoW Versions:**  
+- ElvUI 15.14+ compatibility  
+- Adjusted Nameplate spacing for the new HitTestInset changes (OverlapV 1.8, Click height 22)  
+- Removed Plater from the installer (Still available via /lucky config)  
+- Updated Details profile  
+- Updated the installer steps  
+
+**Retail:**  
+- Added option to select a different ElvUI focus nameplate texture  
+- Fixed an error regarding the new double-clicking to signup option  
+- Updated Northern Sky Raid Tools profile (For next release or alpha version users)  
+- Updated PermoksAccountManager profile  
+- Updated SkironCooldownManager profile (Requires alpha)  
+
+## Version 4.13 [13-05-2026]
 
 **All WoW Versions:**  
 - Added option to disable and hide the UI Errors Frame to /lucky config